+++ bioprov/__init__.py
@@ -2,7 +2,7 @@ __author__ = "Vini Salazar"
 __license__ = "MIT"
 __maintainer__ = "Vini Salazar"
 __url__ = "https://github.com/vinisalazar/bioprov"
-__version__ = "0.1.12"
+__version__ = "0.1.13"
 
 
 """
@@ -11,7 +11,7 @@ Init module for package bioprov.
 Inherits objects from the src/ package.
 """
 
-from .src.config import Config, config, EnvProv
+from .src.config import config, EnvProv, BioProvDB
 from .src.files import File, SeqFile
 from .src.main import (
     Program,